wakeup: use seq_open() to show wakeup source
MIUI-1149480

single_open() interface requires that the whole output must fit into a single buffer. This will lead to timeout when system memory is not in a good situation.

This patch use seq_open() to show wakeup stats. This method need only one page, so timeout will not be observed.

Change-Id: I59e29a19eb91d5e0f92a9bdb6f23857f57815e38
Signed-off-by: Ganesh Mahendran <opensource.ganesh@gmail.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: cuiyangpei <cuiyangpei@xiaomi.com>
Signed-off-by: azrim <mirzaspc@gmail.com>