sched/fair: Prevent tick path active migration to the same CPU
The misfit task running on the lower capacity CPUs are migrated
to higher capacity CPUs from the tick path. It is not guaranteed
that find_energy_efficient_cpu() returns a higher capacity CPU
for the tick path active migration. It may return the previous
CPU or another CPU that has the same capacity as the previous CPU.
There is an explicit check in check_for_migration() to allow the
migration from lower capacity CPU to higher capacity CPU. However
there is a remote possibility of previous CPU capacity being changed
at the same time and we end up queueing the active migration work
on the previous CPU. This results in a kernel panic later when the
migration work actually runs. Prevent this by explicitly checking
the new_cpu against the previous CPU.

Change-Id: I20b25799f388ff609bff7390cd280cf17790adb7
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: azrim <mirzaspc@gmail.com>