usb: dwc3: dwc3-msm: Avoid overriding DBM_GEN_CFG register
The current register write to DBM_GEN_CFG to update the current speed
is overwriting the entire register.  Avoid this, as this also
disables ZLT support on DBM.  During validation of QDSS, a QDSS data
stall was seen, as the QDSS BAM generated a ZLT, while DBM did not
enable it.  DBM ZLT support is recommended to be enabled in the HPG.

Change-Id: Id8b2fc8a81d98854c55005189dcd2c7ae33079f7
Signed-off-by: Wesley Cheng <wcheng@codeaurora.org>
Signed-off-by: Mayank Rana <mrana@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: azrim <mirzaspc@gmail.com>