[BACKPORT]msm: kgsl: Return a valid entry in case of secure GPU address
Currently kgsl_sharedmem_find always returns NULL in case of a secure GPU
address as the address would not be part of process pagetable. This can
lead to not freeing secure memory as part of kgsl_ioctl_sharedmem_free.
Fix this by also checking if the address lies in the secure pagetable
range before returning NULL.

Change-Id: I79b5686a8fcaf580fc79b8f528b91c987f9138a9
Signed-off-by: Puranam V G Tejaswi <pvgtejas@codeaurora.org>
Signed-off-by: Forenche <prahul2003@gmail.com>
Signed-off-by: azrim <mirzaspc@gmail.com>