genirq: Fix spurious "handler enabled interrupts" warning on RT
The SDE IRQ handler routes interrupts itself using generic_handle_irq(),
which coughs up a warning on RT since the SDE IRQ is threaded and therefore
IRQs aren't disabled in the handler. Aside from the warning spam, this is
problematic because the warning unevenly disables IRQs thinking that IRQs
are supposed to be disabled to begin with, which results in all sorts of
mayhem. To fix it, check if IRQs were actually disabled prior to running a
given handler instead of assuming it by default.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: azrim <mirzaspc@gmail.com>