selinux: Make AVC safe on RT
The Access Vector Cache can be invoked from many unpredictable atomic
contexts, such as from printk's kptr_restrict check and from
sched_setscheduler(). Since page allocations may always sleep on RT
regardless of the GFP flags used, we must avoid the allocations entirely
when running in atomic context. Adding a check for this to avc_alloc_node()
covers all the atomic memory allocations in AVC.

In addition, convert notif_lock to a raw spin lock so that it won't sleep,
and move the printk() out from the under the lock while we're at it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: azrim <mirzaspc@gmail.com>