printk: Allow logs to be flushed from IRQ context during a panic on RT
Once a panic happens, scheduling no longer matters and all we care about is
getting debug information. As it stands, crucial panic messages are often
lost on RT because printk refuses to flush the log buffers from IRQ/NMI
context, which is a nightmare for debugging. Fix it by allowing flushes
when oops_in_progress is set.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: azrim <mirzaspc@gmail.com>